QA Update (#529)
* merging dev up to qa (#511) (#512)

* add Dependabot secrets.ACCESS_TOKEN (#454)

* add Dependabot secrets.ACCESS_TOKEN
* revert PR #415

* Bump testcontainers library versions to 1.17.4 (#455)

* Fix for #454

* Second fix for #454

* Revert attempted fixes for dependabot.yml

* Bump testcontainers from 1.17.3 to 1.17.4 (#456)

* Bump testcontainers from 1.17.3 to 1.17.4

Bumps [testcontainers](https://github.com/testcontainers/testcontainers-java) from 1.17.3 to 1.17.4.
- [Release notes](https://github.com/testcontainers/testcontainers-java/releases)
- [Changelog](https://github.com/testcontainers/testcontainers-java/blob/main/CHANGELOG.md)
- [Commits](https://github.com/testcontainers/testcontainers-java/compare/1.17.3...1.17.4)

---
updated-dependencies:
- dependency-name: org.testcontainers:testcontainers
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* use $testcontainers_version


* Bump camel_version to 3.19.0 (#458)

* Python in Gradle (#411)

* test

* python working

* isort fixes

* fix vulnerability

* fix vulnerability

* separate build file

* add python to gradle build

* requirements


* Update Gatling's scala-library version (#461)

* Update Zinc's scala-library version (#462)

* MCP-1918:  Controller shell for MAS Request  (#445)

* MCP-1918: Controller, request/response, validation test

* MCP-1918: fixed checkstyle warnings

* MCP-1918: lowercase to camelcase

* MCP-1918: lowercase to camelcase

* MCP-1918: checkstyle warning

* MCP-1935: Use common domain module to avoid duplicate objects (#463)

* MCP-1935: extracted common domain

* MCP-1935: extracted common domain

* Update Gradle artifact group name to gov.va.vro (#465)

* group=gov.va.vro
* update taskinfo_plugin to 2.0.0
* remove group name from service-data-access

* Disable Trivy's image scan

* MCP-1937: Upgrade snakeyaml (#467)

* MCP-1937: Upgrade snakeyaml

* MCP-1937: upgarde jackson

* MCP-1937: upgrade jackson

* Rename starter-java.jar to vro-app.jar (#469)

* Feature/mcp 1929 remove gradle warnings (#466)

* removw jcenter

* merge domain enums

* replace deprecated method


* push config to qa (#474)

* Github Action Cleanup (#471)

Action Cleanup


* Use unshaded ObjectMapper (#476)

* Bump okhttp3 to 4.10.0 (#477)

* Bump all snakeyaml to 1.33 (#478)

* revert jackson - fixes swagger ui (#473)

* revert jackson - fixes swagger ui

* revert jackson - fixes swagger ui

* Mcp 1923 - Update Security Configuration to avoid using deprecated components. (#464)

* Replace deprecated WebSecurityConfigurAdapter. See MCP-1932. In addition, add log for access attempts using wrong keys as required by VA SD.

* Remove unnecessary code.

* Use SCM for versioning (#484)

* remove service-data-access version
* configure axion-release
* subprojects
* fileUpdate hook
* use defaults
* Release version: 1.0.0

* Aqua Trivy action: don't create/scan images (#487)

* don't create images
* don't upload results for now
* output table to console

* Use Alpine-based Docker images (#483)

* use Alpine-based Docker images

* make consistent and more readable

* pin curl version

* Bump Alpine's expat to 2.4.9-r0 (#488)

* Clean up build.gradle files with dependency constraints (#485)

* add nebula.lint

* api: reorder dep; clarify comment

* clean up api

* clean up app

* clean up domain

* clean up controller

* clean up service/*

* reorder project deps

* remove unneeded deps

* consolidate annotationBom and checkstyleRules


* Delint Gradle dependencies; restore health endpoint (#489)

* add more lint rules
* testRuntimeOnly h2
* runtimeOnly jackson-datatype-jsr310
* restore health endpoint and other va.starter deps
* add gradle-lint.sh

* MCP-1924 Python Consumer Updates (#459)

* consumer updates

* lint

* logic cleanup

* Add log stack trace


* Mcp 1951 - Fix integration test in service-data-access (#480)

* Replace deprecated WebSecurityConfigurAdapter. See MCP-1932. In addition, add log for access attempts using wrong keys as required by VA SD.

* Remove unnecessary code.

* Fix integration test in service-data-access. See MCP-1951

* Fix spotless check error.


* Bump hapi-fhir-base from 6.1.1 to 6.1.2 (#498)

Bumps hapi-fhir-base from 6.1.1 to 6.1.2.

---
updated-dependencies:
- dependency-name: ca.uhn.hapi.fhir:hapi-fhir-base
  dependency-type: direct:production
  update-type: version-update:semver-patch
...



* Bump hapi-fhir-server-openapi from 6.1.1 to 6.1.2 (#496)

Bumps hapi-fhir-server-openapi from 6.1.1 to 6.1.2.

---
updated-dependencies:
- dependency-name: ca.uhn.hapi.fhir:hapi-fhir-server-openapi
  dependency-type: direct:production
  update-type: version-update:semver-patch
...



* Python Requirements Update (#503)

* Requirements Update

* Version Update

* [API-1234] take out extra requirements


* MCP-1920: MAS integration polling queue (#468)

* MCP-1920: Start off

* MCP-1920: moved to common domain

* MCP-1920: setup camel route

* MCP-1920: Fixed imports

* MCP-1920: queue connections

* MCP-1920: polling service

* MCP-1920: asynchronous polling

* MCP-1920: configure delays

* MCP-1920: re-added missing dependency

* MCP-1920: payload definition changed

* MCP-1920: payload definition changed

* MCP-1920: code review comments

* MCP-1920: code review comments



* update sent-env to >> $GITHUB_ENV (#521)

* Config/set env update (#522)

* update sent-env to >> $GITHUB_ENV

* updating action versions

* Feature/mcp 1962 disable endpoints (#524)

* separate end points to be disabled

* separate controllers to be disabled

* separate controller tests

* code cleaning

* code cleaning

* remove dev controller from qa

* remove unused


* Adjust project dependencies (#526)

* app is not dep on microservice containers
* don't rebuild db-init docker image
* remove old reference to gov.va.starter.example.persistence.model
* use image cache
* Spring's RabbitHealthIndicator needs a connection to RabbitMQ
