+++ .github/actions/install-java-tools/action.yml
@@ -7,7 +7,7 @@ runs:
     # http://man7.org/linux/man-pages/man1/date.1.html
     - name: Get Date
       id: get-date
-      run: echo "::set-output name=date::$(/bin/date -u "+%Y%m%d")"
+      run: echo "name=date::$(/bin/date -u "+%Y%m%d")" >> $GITHUB_ENV
       shell: bash
 
     - name: Restore hadolint
