+++ .github/actions/setup-vro/action.yml
@@ -6,7 +6,7 @@ runs:
   using: composite
   steps:
     - name: "Install Java"
-      uses: actions/setup-java@v3
+      uses: actions/setup-java@v3.5.1
       with:
         distribution: temurin
         java-version: '17'
@@ -15,7 +15,7 @@ runs:
     - uses: ./.github/actions/install-java-tools
 
     - name: "Install Python"
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v4.3.0
       with:
         python-version: "3.9"
         cache: "pip"
