+++ app/build.gradle
@@ -71,6 +71,10 @@ tasks.dockerPrepare.configure {
     dependsOn(tasks.bootJar)
 }
 
+docker {
+    noCache false
+}
+
 bootJar {
     manifest {
         attributes(
@@ -119,6 +123,7 @@ tasks.named('lintDockerfile').configure {
     ext.targets = ["src/docker/Dockerfile"]
 }
 
-tasks.named('dockerComposeUp').configure {
-    dependsOn ':db-init:docker'
-}
+//Needed? this rebuilds the db-init docker image every time dockerComposeUp is run
+//tasks.named('dockerComposeUp').configure {
+//    dependsOn ':db-init:docker'
+//}
